Make the definition of we explicit throughout the document.
It was brought up in the meeting of the current authors that "we" in our document was not
well defined to our audience.  They may be confused as to who "we" is referring to, and if
it includes them or not.  This commit makes that option to include themselves explicit.