First draft of a basic explanation of ethics (#3)
Previously, we set out to define what people should and should not do.  However,
it became apparent that we forgot to first explain how a person should come to conclusions
about what they should, or should not do.  This is necessary to allow people freedom, but
also to justify which maxims go into this book.

If we care about our posterity, informing them how to decide how to act, is more important
that teaching them how to act.  Although both things are important for those children whom
are hard-headed and slow to learn.  If they know why, eventually the how will be remembered.