Complete reworking of the ordering and content
Currently, the Book is organized in a rather heady manner. However, the goal of the work is to improve as many people's lives as possible. Therefor, all the current content have been moved to an appendix, and the introduction rewritten.  The book should be organized from most important, to least important, content for people to know.  Unfortunately, this means people who believe themselves to be wise may never start reading it -- but there will even be content for them later in the book.  Hopefully, this organization has the most impact on the most people.