Add a note about the divinity of mankind.
Men and women are made in the image of God.  As such, they have intrinsic
value.  That is to say, they are valuable simply due to their existance.
All functioning legal systems are predicated upon this idea.

If humans do not have intrinsic value, then there is no sense in having
any moral code at all. You may as do whatever you have the power to do --
but no-one can survive for long in a world ruled by pure strength.

The acknowledgement of the intrinsic value of the individual puts
limitations on the depth to which civilization can plunge itself.